get rid of old story [ui skip]
